⚒ Permissions, esthétique des embeds, ajout d'un dev mode et relecture

+++ main.py
@@ -3,11 +3,14 @@ Classe principale du bot.
 """
 # Imports
 import discord
+from discord.ext import commands
 import config
 import database as db
-
+from pytz import timezone
 
 # Initisalisation du bot
+
+
 class AClient(discord.Client):
     """
     Discord client
@@ -27,6 +30,12 @@ class AClient(discord.Client):
             self.add_view(TickerLauncher())
             self.added = True
         print(f"🤖 Connexion réussie : {self.user}.")
+        # Logs
+        embed = discord.Embed(title="🟢 Le bot est en ligne !",
+                              color=discord.Colour.dark_green())
+        if not config.get("DEV_MODE"):
+            channelLog = client.get_channel(1068629560209440780)
+            await channelLog.send(embed=embed)
 
     async def setup_hook(self) -> None:
         self.add_view(MainView())
@@ -43,10 +52,9 @@ tree = discord.app_commands.CommandTree(client)
 feedback_listen = dict()
 
 
-# Système de tickets
 class TickerLauncher(discord.ui.View):
     """
-    c'est un objet contenant 1 bouton avec l'évenement création ticket
+    Objet contenant 1 bouton avec l'évenement création ticket
     """
 
     def __init__(self) -> None:
@@ -65,6 +73,11 @@ class TickerLauncher(discord.ui.View):
             await interaction.response.send_message(
                 f"🇫🇷 Vous avez déjà un ticket ouvert à {ticket.mention}!\n\n🇬🇧🇺🇸 You already have a ticket open at {ticket.mention}!",
                 ephemeral=True)
+        elif interaction.user.guild_permissions.send_messages is False:
+            # L'utilisateur a été mute
+            await interaction.response.send_message(
+                f"🇫🇷 Vous avez été mute, vous ne pouvez pas ouvrir de ticket!\n\n🇬🇧🇺🇸 You have been muted, you cannot open a ticket!",
+                ephemeral=True)
         else:
             # Création du ticket
             overwrites = {
@@ -75,13 +88,6 @@ class TickerLauncher(discord.ui.View):
                 interaction.guild.me: discord.PermissionOverwrite(view_channel=True, send_messages=True,
                                                                   read_message_history=True)
             }
-
-            # Log ouverture ticket
-            channelLog = client.get_channel(1068629560209440780)
-
-            embed = discord.Embed(title="un ticket vien de se crée ",
-                                 description=f" c'est le ticket de {interaction.user} ", color=discord.Colour.green())
-            await channelLog.send(embed=embed)
 
             # Création du channel
             category = discord.utils.get(
@@ -98,11 +104,25 @@ class TickerLauncher(discord.ui.View):
                 f"🇫🇷 J'ai ouvert un ticket pour vous ici {channel.mention}!\n\n🇬🇧🇺🇸 I've opened a ticket for you at {channel.mention}!",
                 ephemeral=True)
 
+            # Log ouverture ticket
+            channelLog = client.get_channel(1068629560209440780)
 
-# Système de fermeture de tickets
+            # Date conversion et formatage
+            date = interaction.created_at
+            date = date.astimezone(tz=timezone('Europe/Paris'))
+            date = date.strftime("%d/%m/%Y à %H:%M:%S")
+
+            embed = discord.Embed(title="🎫 TICKET CREE",
+                                  description=f"""
+                                  **Nom du channel :** {channel.name}
+                                  \n**Crée par :** {interaction.user.mention}
+                                  \n**Date de création :** {date}""", color=discord.Colour.green())
+            await channelLog.send(embed=embed)
+
+
 class ConfirmView(discord.ui.View):
     """
-    objet permetant de confirmer la fermeture du ticket , il comporte 1 bouton
+    Objet contenant 1 bouton avec l'évenement confirmation fermeture ticket
     """
 
     def __init__(self) -> None:
@@ -110,26 +130,42 @@ class ConfirmView(discord.ui.View):
 
     @discord.ui.button(label="Confirm", style=discord.ButtonStyle.red, custom_id="confirm")
     async def confirm_button(self, interaction: discord.Interaction, button: discord.ui.Button):
+        if interaction.user.guild_permissions.manage_channels is False:
+            await interaction.response.send_message(
+                f"🇫🇷 Vous n'avez pas la permission de fermer ce ticket!\n\n🇬🇧🇺🇸 You don't have permission to close this ticket!",
+                ephemeral=True)
+            return
         try:
             # Log fermeture ticket
             channelLog = client.get_channel(1068629560209440780)
+            # Fetch des users ayant parlé dans le channel
+            users = [message.author.mention async for message in interaction.channel.history(limit=200)]
+            users = list(set(users))  # Suppression des doublons
+            users = ", ".join(users)  # Conversion en string
 
-            embed = discord.Embed(title="un ticket vien de se fermer ",
-                                 description=f" c'est le ticket de {interaction.user} ", color=discord.Colour.red())
+            # Date conversion et formatage
+            date = interaction.created_at
+            date = date.astimezone(tz=timezone('Europe/Paris'))
+            date = date.strftime("%d/%m/%Y à %H:%M:%S")
+
+            embed = discord.Embed(title="🎫 TICKET FERME",
+                                  description=f"""**Nom du channel :** {interaction.channel.name}
+                                    \n**Fermé par :** {interaction.user.mention}
+                                    \n **Utilisateurs ayant parlé dans le ticket :** {users}
+                                    \n**Date de fermeture :** {date}""", color=discord.Colour.red())
             await channelLog.send(embed=embed)
-
 
             # delet channel
             await interaction.channel.delete()
         except:
             await interaction.response.send_message(
-                "🇫🇷 Impossible de supprimer le channel. Merci de vérifier que je possède la permission MANAGE_CHANNELS.\n\n\n🇬🇧🇺🇸 I can't delete this channel. Please check that i have the MANAGE_CHANNELS permission.",
+                "🇫🇷 Impossible de supprimer le channel. Merci de vérifier que je possède la permission MANAGE_CHANNELS.\n\n🇬🇧🇺🇸 I can't delete this channel. Please check that i have the MANAGE_CHANNELS permission.",
                 ephemeral=True)
 
 
 class MainView(discord.ui.View):
     """
-    objet principal qui gère le ticket une fois ouvert il comporte 3 bouton
+    Objet contenant 2 boutons avec les évenements de fermeture et d'archivage de tickets    
     """
 
     def __init__(self) -> None:
@@ -142,14 +178,13 @@ class MainView(discord.ui.View):
 
     @discord.ui.button(label="Archive", custom_id="ticket_archive", style=discord.ButtonStyle.blurple)
     async def archive(self, interaction: discord.Interaction, button: discord.ui.Button):
-        msg="🇫🇷 Voulez-vous vraiment archiver ce ticket ?\n\n🇬🇧🇺🇸 Are you sure you want to archive this ticket ?"
+        msg = "🇫🇷 Voulez-vous vraiment archiver ce ticket ?\n\n🇬🇧🇺🇸 Are you sure you want to archive this ticket ?"
         await interaction.response.send_message(msg, view=ArchiveConfirm(), ephemeral=True)
 
 
-# Système d'archivation de tickets
 class ArchiveConfirm(discord.ui.View):
     """
-    sert a confirmer l'archivation , cet objet comporte 1 bouton
+    Objet contenant 1 bouton avec l'évenement confirmation archivage ticket
     """
 
     def __init__(self) -> None:
@@ -157,6 +192,11 @@ class ArchiveConfirm(discord.ui.View):
 
     @discord.ui.button(label="Confirm", style=discord.ButtonStyle.red, custom_id="confirm")
     async def confirm_button(self, interaction: discord.Interaction, button: discord.ui.Button):
+        if interaction.user.guild_permissions.manage_channels is False:
+            await interaction.response.send_message(
+                f"🇫🇷 Vous n'avez pas la permission d'archiver ce ticket!\n\n🇬🇧🇺🇸 You don't have permission to archive this ticket!",
+                ephemeral=True)
+            return
         try:
             category = discord.utils.get(
                 interaction.guild.categories, id=1061049218569084948)
@@ -169,31 +209,45 @@ class ArchiveConfirm(discord.ui.View):
 
             await channel.edit(category=category, name=channel.name + "-archived")
             await interaction.response.defer()
-            
+
             # Log archive ticket
             channelLog = client.get_channel(1068629560209440780)
 
-            embed = discord.Embed(title="un ticket vien de s'archiver ",
-                                 description=f" archiver par  {interaction.user} ", color=discord.Colour.blue())
+            # Log fermeture ticket
+            channelLog = client.get_channel(1068629560209440780)
+            # Fetch des users ayant parlé dans le channel
+            users = [message.author.mention async for message in interaction.channel.history(limit=200)]
+            users = list(set(users))  # Suppression des doublons
+            users = ", ".join(users)  # Conversion en string
+
+            # Date conversion et formatage
+            date = interaction.created_at
+            date = date.astimezone(tz=timezone('Europe/Paris'))
+            date = date.strftime("%d/%m/%Y à %H:%M:%S")
+
+            embed = discord.Embed(title="🎫 TICKET ARCHIVE",
+                                  description=f"""**Channel :** {interaction.channel.mention}
+                                    \n**Fermé par :** {interaction.user.mention}
+                                    \n **Utilisateurs ayant parlé dans le ticket :** {users}
+                                    \n**Date de fermeture :** {date}""", color=discord.Colour.blurple())
             await channelLog.send(embed=embed)
-        
+
         except:
             await interaction.response.send_message(
                 "🇫🇷 Impossible de déplacer le channel. Merci de vérifier que je possède la permission MANAGE_CHANNELS.\n\n🇬🇧🇺🇸 I can't archive this channel. Please check that i have the MANAGE_CHANNELS permission.",
                 ephemeral=True)
 
 
-# feedback système
 class FeedBack(discord.ui.View):
     """
-    gère le système de feedback , cet objet comporte 6 bouton
+    Objet contenant 6 boutons avec les évenements de feedback
     """
 
     def __init__(self, freelancer=None) -> None:
         super().__init__(timeout=None)
         self.tab = [None, None, None, freelancer]
 
-        # Button star
+    # Button star
     @discord.ui.button(label="⭐️", custom_id="1_star", style=discord.ButtonStyle.blurple)
     async def star_1(self, interaction: discord.Interaction, button: discord.ui.Button):
         self.tab[0] = interaction.user
@@ -212,7 +266,7 @@ class FeedBack(discord.ui.View):
     async def star_3(self, interaction: discord.Interaction, button: discord.ui.Button):
         self.tab[0] = interaction.user
         self.tab[2] = 3
-        await interaction.response.send_message("🇫🇷 Vous avez mis une note de 3 étoiles. \n\n🇬🇧🇺🇸 You have given a 3 stars rating." , ephemeral=True)
+        await interaction.response.send_message("🇫🇷 Vous avez mis une note de 3 étoiles. \n\n🇬🇧🇺🇸 You have given a 3 stars rating.", ephemeral=True)
 
     # Button star
     @discord.ui.button(label="⭐️⭐️⭐️⭐️", custom_id="4_star", style=discord.ButtonStyle.blurple)
@@ -245,6 +299,20 @@ class FeedBack(discord.ui.View):
     async def finish(self, interaction: discord.Interaction, button: discord.ui.Button):
         if not None in self.tab:
             db.add_avis(self.tab[0], self.tab[3], self.tab[1], self.tab[2])
+
+            # Embed
+            feedBackChannel = client.get_channel(
+                1059124591852793916)  # TODO : Mettre le bon id
+            embed = discord.Embed(title="📝 FEEDBACK",
+                                  description=f"""
+            **Client :** {interaction.user.mention}
+            **Freelancer :** {self.tab[3]}
+            **Note :** {self.tab[2]}/5
+            **Commentaire :** {self.tab[1]}
+            """, color=discord.Color.purple())
+            embed.set_thumbnail(url=interaction.user.avatar)
+            await feedBackChannel.send(embed=embed)
+
             await interaction.response.send_message("🇫🇷 **Merci pour votre retour !** 💖 \n\n🇬🇧🇺🇸 **Thank you for your feedback !** 💖")
         else:
             await interaction.response.send_message(
@@ -262,6 +330,7 @@ async def ping(interaction: discord.Interaction):
 
 
 @tree.command(name="test", description="Test dev", guild=discord.Object(id=1046437841447686226))
+@commands.has_permissions(administrator=True) 
 async def test(interaction: discord.Interaction):
     """
     Test dev
@@ -273,9 +342,9 @@ async def test(interaction: discord.Interaction):
     await interaction.response.send_message(embed=embed)
 
 
-# Commandes
 @tree.command(name="clear", description="Retirer des messages d'un channel",
               guild=discord.Object(id=1046437841447686226))
+@commands.has_permissions(manage_channels=True)
 async def self(ctx, amount: int = None):
     """
     Retire des messages d'un channel
@@ -296,10 +365,13 @@ async def self(ctx, amount: int = None):
             await ctx.channel.purge(limit=amount)
 
 
-# Commandes
 @tree.command(name="ticket", guild=discord.Object(id=1046437841447686226),
               description="Lance le système de ticket en affichant le message avec la réaction")
+@commands.has_permissions(administrator=True)
 async def ticketing(interaction: discord.Interaction):
+    """
+    Lance le système de ticket en affichant le message avec la réaction
+    """
     embed = discord.Embed(
         title="🇬🇧🇺🇸 If you need support or want to order, click the button below to create a ticket ! \n\n🇫🇷 Si vous avez besoin d'aide ou que vous souhaitez passer commande, clickez sur le bouton ci-dessous pour créer un ticket !",
         color=discord.Colour.blue())
@@ -307,21 +379,26 @@ async def ticketing(interaction: discord.Interaction):
     await interaction.response.send_message("✅ Système de ticket lancé avec succès !", ephemeral=True)
 
 
-# Commandes
 @tree.command(name="close", guild=discord.Object(id=1046437841447686226), description="Ferme le ticket")
+@commands.has_permissions(manage_channels=True)
 async def close(interaction: discord.Interaction):
+    """
+    Ferme le ticket
+    """
     if "ticket-for-" in interaction.channel.name:
-        msg="🇬🇧🇺🇸 Are you sure you want to close this ticket ?\n\n🇫🇷Voulez-vous vraiment fermer ce ticket ?"
+        msg = "🇬🇧🇺🇸 Are you sure you want to close this ticket ?\n\n🇫🇷Voulez-vous vraiment fermer ce ticket ?"
         await interaction.response.send_message(msg, view=ConfirmView(), ephemeral=True)
     else:
         await interaction.response.send_message(
             "🇬🇧🇺🇸 This channel isn't a ticket !\n\n🇫🇷Ce channel n'est pas un ticket !", ephemeral=True)
 
 
-# Commandes
 @tree.command(name="add", guild=discord.Object(id=1046437841447686226), description="Ajoute un utilisateur au ticket")
 @discord.app_commands.describe(user="L'utilisateur à ajouter au ticket")
 async def add(interaction: discord.Interaction, user: discord.Member):
+    """
+    Ajoute un utilisateur au ticket
+    """
     if "ticket-for-" in interaction.channel.name:
         await interaction.channel.set_permissions(user, view_channel=True, send_messages=True, attach_files=True,
                                                   embed_links=True)
@@ -331,35 +408,62 @@ async def add(interaction: discord.Interaction, user: discord.Member):
             "🇬🇧🇺🇸 This channel isn't a ticket !\n\n🇫🇷 Ce channel n'est pas un ticket !", ephemeral=True)
 
 
-# Commandes
 @tree.command(name="feedback", guild=discord.Object(id=1046437841447686226), description="Lance le système de feedback")
 async def launchefeedback(interaction: discord.Interaction):
+    """
+    Lance le système de feedback en affichant le message avec les réactions
+    """
     embed = discord.Embed(title="🌟 FEEDBACK", description="🇫🇷 Afin d'avoir un retour clair sur notre service, nous vous invitons à ajouter un commentaire et une note à E-shop en utilisant les boutons ci-dessous ! Cela ne prendra que quelques minutes.\n\n🇬🇧🇺🇸 To have a honnest feedback on our service, we invite you to add a comment and a rating to E-shop using the buttons below ! This will only take a few minutes.",
                           color=discord.Colour.blue())
     await interaction.channel.send(embed=embed,
                                    view=FeedBack(f"{interaction.user.name} - {interaction.user.discriminator}"))
     await interaction.response.send_message("✅ Système de feedback lancé avec succès !", ephemeral=True)
 
-# Commandes
 
-
-@tree.command(name="ptsfidelite", guild=discord.Object(id=1046437841447686226), description="fidelise le client")
+@tree.command(name="points", guild=discord.Object(id=1046437841447686226), description="fidelise le client")
 @discord.app_commands.choices(param=[
     discord.app_commands.Choice(name="reset", value='reset'),
     discord.app_commands.Choice(name="ajouter", value='add'),
     discord.app_commands.Choice(name="retirer", value='remove'),
 ])
 async def ptsfidelite(interaction: discord.Interaction, acheteur: discord.Member, param: str, nbre_point: int):
+    """
+    Modifie les points de fidélité d'un client
+    """
     if param == 'reset':
         db.reset_client_points(acheteur)
+        await interaction.response.send_message(f"✅ Points de fidélité de {acheteur} réinitialisés avec succès !", ephemeral=True)
 
     elif param == 'add':
         db.add_client_points(acheteur, nbre_point)
+        await interaction.response.send_message(f"✅ {nbre_point} points de fidélité ajoutés à {acheteur} avec succès !", ephemeral=True)
 
     elif param == 'remove':
         db.remove_client_points(acheteur, nbre_point)
+        await interaction.response.send_message(f"✅ {nbre_point} points de fidélité retirés à {acheteur} avec succès !", ephemeral=True)
 
 
+@tree.error
+@test.error
+async def error_handler(interaction: discord.Interaction, error):
+    if isinstance(error, discord.errors.Forbidden):
+        await interaction.response.send_message("❌ Je n'ai pas les permissions nécessaires pour effectuer cette action !", ephemeral=True)
+    elif isinstance(error, discord.errors.HTTPException):
+        await interaction.response.send_message("❌ Une erreur est survenue lors de l'envoi du message !", ephemeral=True)
+    elif isinstance(error, discord.errors.NotFound):
+        await interaction.response.send_message("❌ Une erreur est survenue lors de l'envoi du message !", ephemeral=True)
+    elif isinstance(error, commands.errors.MissingRequiredArgument):
+        await interaction.response.send_message("❌ Vous n'avez pas renseigné tous les arguments nécessaires !", ephemeral=True)
+    elif isinstance(error, commands.errors.BadArgument):
+        await interaction.response.send_message("❌ Vous n'avez pas renseigné un argument correct !", ephemeral=True)
+    elif isinstance(error, commands.errors.CommandInvokeError):
+        await interaction.response.send_message("❌ Une erreur est survenue lors de l'envoi du message !", ephemeral=True)
+    elif isinstance(error, commands.errors.CommandNotFound):
+        await interaction.response.send_message("❌ Cette commande n'existe pas !", ephemeral=True)
+    elif isinstance(error, commands.errors.CheckFailure) or isinstance(error, commands.errors.MissingPermissions):
+        await interaction.response.send_message("❌ Vous n'avez pas les permissions nécessaires pour effectuer cette action !", ephemeral=True)
+
+# Message de bienvenue
 @client.event
 async def on_member_join(member):
     channel = member.guild.system_channel
@@ -368,6 +472,7 @@ async def on_member_join(member):
     embed.set_thumbnail(url=f"{member.display_avatar}")
     await channel.send(embed=embed)
 
+# Commentaires pour le feedback
 @client.event
 async def on_message(message):
     if message.channel not in feedback_listen:
